+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Handle timezone offset changes on remote FAT filesystems
 - Drop reading the old state storage format
+- Do not mention my forked prim-ftpd, PR-s got merged in the original repo
 
 ## 0.6.0