+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Drop reading the old state storage format
 - Do not mention my forked prim-ftpd, PR-s got merged in the original repo
 - Remove --valid-chars option
+- Add documentation about using ssh-agent for passphrases
 
 ## 0.6.0